requirements.txt
appnope==0.1.4
asttokens==3.0.0
backcall==0.2.0
blinker==1.8.2
click==8.1.8
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.57.0
fsspec==2025.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.7
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.4.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
